Moved classes to structures package, and added equals/hashCode/toString
