+++ algorithms/src/main/com/ictools/algorithms/graph/structures/Node.java
@@ -1,4 +1,4 @@
-package com.ictools.algorithms.graph;
+package com.ictools.algorithms.graph.structures;
 
 public interface Node {
     long getIdentifier();
